Rename flow.ts to vexflow.ts.
