+++ entry/vexflow-bravura.ts
@@ -5,7 +5,7 @@
 // vexflow-bravura.ts is the entry point for vexflow-bravura.js.
 // This version bundles the Bravura music engraving font and the Academico text font.
 
-import { Flow } from '../src/flow';
+import { Flow } from '../src/vexflow';
 import { Font } from '../src/font';
 import { Academico } from '../src/fonts/academico';
 import { AcademicoBold } from '../src/fonts/academicobold';
