+++ entry/vexflow-debug.ts
@@ -12,7 +12,7 @@
 // It bundles the same fonts as vexflow.js.
 // Other music/text fonts need to be loaded at runtime during page load. See VexFlow.loadFonts().
 
-import { Flow } from '../src/flow';
+import { Flow } from '../src/vexflow';
 import { Font } from '../src/font';
 import { Academico } from '../src/fonts/academico';
 import { AcademicoBold } from '../src/fonts/academicobold';
