+++ entry/vexflow.ts
@@ -5,7 +5,7 @@
 
 // The webpack config inside Gruntfile.js sets the webpack mode to 'production' to produce a minified build.
 
-import { Flow } from '../src/flow';
+import { Flow } from '../src/vexflow';
 import { Font } from '../src/font';
 import { Academico } from '../src/fonts/academico';
 import { AcademicoBold } from '../src/fonts/academicobold';
