+++ tests/notehead_tests.ts
@@ -10,7 +10,7 @@
 
 import { TestOptions, VexFlowTests } from './vexflow_test_helpers';
 
-import { Flow } from '../src/flow';
+import { Flow } from '../src/vexflow';
 import { Formatter } from '../src/formatter';
 import { NoteHead } from '../src/notehead';
 import { RenderContext } from '../src/rendercontext';
