+++ tests/tabtie_tests.ts
@@ -6,7 +6,7 @@
 import { TestOptions, VexFlowTests } from './vexflow_test_helpers';
 
 import { Annotation } from '../src/annotation';
-import { Flow } from '../src/flow';
+import { Flow } from '../src/vexflow';
 import { Formatter } from '../src/formatter';
 import { Note } from '../src/note';
 import { RenderContext } from '../src/rendercontext';
