+++ tests/voice_tests.ts
@@ -5,7 +5,7 @@
 
 import { TestOptions, VexFlowTests } from './vexflow_test_helpers';
 
-import { Flow } from '../src/flow';
+import { Flow } from '../src/vexflow';
 import { Formatter } from '../src/formatter';
 import { ContextBuilder } from '../src/renderer';
 import { Stave } from '../src/stave';
